+++ packages/intl-phone-input/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './select-field';
+export * from './flag-icon';
+export * from './select';